+++ RELEASE_NOTES.md
@@ -2,6 +2,7 @@
 - Added GrayAlpha pixel format constants
 - Added ToString() override for PixFormat
 - Added PixFileFormat.Webp and PixWebpSaveParams
+- Added more known image file extensions
 
 ### 5.3.10
 - [IO] Added SymMap constructor overload